Refactored the project to allow the DataTable to be used with any data type.
Added Blazor WASM and Blazor Server example projects, that use the FilterTable project to create a log viewer, accessing log entries from a local database, by calling the web service in the LogEntryDataAccess project.

The FilterTable, FilterDataAccess and FilterTypes projects can be used without the rest of the example projects or the projects can be used as a basis to retrieve other data types.